+++ crypto/keys/bip39/wordcodec_test.go
@@ -3,13 +3,13 @@ package bip39
 import (
 	"testing"
 
-	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestWordCodec_NewMnemonic(t *testing.T) {
 	_, err := NewMnemonic(FundRaiser)
-	assert.NoError(t, err, "unexpected error generating fundraiser mnemonic")
+	require.NoError(t, err, "unexpected error generating fundraiser mnemonic")
 
 	_, err = NewMnemonic(FreshKey)
-	assert.NoError(t, err, "unexpected error generating new 24-word mnemonic")
+	require.NoError(t, err, "unexpected error generating new 24-word mnemonic")
 }