+++ x/ibc/ibc_test.go
@@ -3,7 +3,7 @@ package ibc
 import (
 	"testing"
 
-	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
 	abci "github.com/tendermint/tendermint/abci/types"
 	"github.com/tendermint/tendermint/crypto"
@@ -72,8 +72,8 @@ func TestIBC(t *testing.T) {
 	mycoins := sdk.Coins{sdk.NewCoin("mycoin", 10)}
 
 	coins, _, err := ck.AddCoins(ctx, src, mycoins)
-	assert.Nil(t, err)
-	assert.Equal(t, mycoins, coins)
+	require.Nil(t, err)
+	require.Equal(t, mycoins, coins)
 
 	ibcm := NewMapper(cdc, key, DefaultCodespace)
 	h := NewHandler(ibcm, ck)
@@ -93,23 +93,23 @@ func TestIBC(t *testing.T) {
 	var igs int64
 
 	egl = ibcm.getEgressLength(store, chainid)
-	assert.Equal(t, egl, int64(0))
+	require.Equal(t, egl, int64(0))
 
 	msg = IBCTransferMsg{
 		IBCPacket: packet,
 	}
 	res = h(ctx, msg)
-	assert.True(t, res.IsOK())
+	require.True(t, res.IsOK())
 
 	coins, err = getCoins(ck, ctx, src)
-	assert.Nil(t, err)
-	assert.Equal(t, zero, coins)
+	require.Nil(t, err)
+	require.Equal(t, zero, coins)
 
 	egl = ibcm.getEgressLength(store, chainid)
-	assert.Equal(t, egl, int64(1))
+	require.Equal(t, egl, int64(1))
 
 	igs = ibcm.GetIngressSequence(ctx, chainid)
-	assert.Equal(t, igs, int64(0))
+	require.Equal(t, igs, int64(0))
 
 	msg = IBCReceiveMsg{
 		IBCPacket: packet,
@@ -117,18 +117,18 @@ func TestIBC(t *testing.T) {
 		Sequence:  0,
 	}
 	res = h(ctx, msg)
-	assert.True(t, res.IsOK())
+	require.True(t, res.IsOK())
 
 	coins, err = getCoins(ck, ctx, dest)
-	assert.Nil(t, err)
-	assert.Equal(t, mycoins, coins)
+	require.Nil(t, err)
+	require.Equal(t, mycoins, coins)
 
 	igs = ibcm.GetIngressSequence(ctx, chainid)
-	assert.Equal(t, igs, int64(1))
+	require.Equal(t, igs, int64(1))
 
 	res = h(ctx, msg)
-	assert.False(t, res.IsOK())
+	require.False(t, res.IsOK())
 
 	igs = ibcm.GetIngressSequence(ctx, chainid)
-	assert.Equal(t, igs, int64(1))
+	require.Equal(t, igs, int64(1))
 }