+++ x/slashing/tick_test.go
@@ -46,8 +46,8 @@ func TestBeginBlocker(t *testing.T) {
 
 	height := int64(0)
 
-	// for 50 blocks, mark the validator as having signed
-	for ; height < 50; height++ {
+	// for 1000 blocks, mark the validator as having signed
+	for ; height < SignedBlocksWindow; height++ {
 		ctx = ctx.WithBlockHeight(height)
 		req = abci.RequestBeginBlock{
 			Validators: []abci.SigningValidator{{
@@ -58,8 +58,8 @@ func TestBeginBlocker(t *testing.T) {
 		BeginBlocker(ctx, req, keeper)
 	}
 
-	// for 51 blocks, mark the validator as having not signed
-	for ; height < 102; height++ {
+	// for 500 blocks, mark the validator as having not signed
+	for ; height < ((SignedBlocksWindow * 2) - MinSignedPerWindow + 1); height++ {
 		ctx = ctx.WithBlockHeight(height)
 		req = abci.RequestBeginBlock{
 			Validators: []abci.SigningValidator{{